Add `neighbor_kernel.h`, remove anonymous namespace (#124)
It seems that the PyTorch register / dispatching logic is not fully
reliable, which causes problems when depending on `neighbor.h` from
external libraries. Until this is resolved in a more principled way, we
introduce `neighbor_kernel.h` here for external libraries that wish to
depend directly on `*_kernel` implementations, that avoid that logic.