Add function to merge samplers outputs (#252)
This code belongs to the part of the whole distributed training for PyG.

## Description
Distributed training requires after each layer to merge results between
machines. For later algorithms, it is required that the results be
sorted according to the sampling order. This PR introduces a function
which purpose is to handle merge and sort operations in parallel.

**Other distributed PRs:**
pytorch_geometric DistLoader:
[#7869](https://github.com/pyg-team/pytorch_geometric/pull/7869)
pytorch_geometric DistSampler:
[#7974](https://github.com/pyg-team/pytorch_geometric/pull/7974)
pyg-lib: [#246](https://github.com/pyg-team/pyg-lib/pull/246)

---------

Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>